added check_if_task_exists method to todoist.
Moved modify_task to end of clickup webhook.
This function call checked if the task existed in todoist so a separate
method has been made.

+++ main.py
@@ -112,44 +112,45 @@ def clickup_webhook_received():
             "task": clickupTask,
         }
         outputData = {"platform": todoist}
+        todoistTaskExists = todoist.check_if_task_exists(clickupTask)
 
-        # Clickup task into todoist next_actions:
-        if clickupRequest["event"] in ["task_updated"]:
-            # Check if valid for next actions
-            # Next action status and (high priority, due date < 1 week, no project.)
-            if (
-                clickupRequest["event"] in ["task_updated"]
-                and clickupTask["status"] == "next action"
-                and (
-                    clickupTask["priority"] < 3
-                    or max_days_diff(clickupTask["due_date"], days=3)
-                    or not clickup.is_subtask(clickupTask)
-                )
-            ):
-                outputData["list"] = "next_actions"
-                if todoistTask == False:
-                    logger.info(f"Adding task to next actions list.")
-                    todoistTask = move_task(inputData, outputData, deleteTask=False)
-                    clickup.add_todoist_id(clickupTask, todoistTask["todoist_id"])
-                else:
-                    logger.info(f"Task is already in next actions list.")
-            elif (
-                clickupRequest["event"] in ["task_updated"]
-                and todoistTask != False
-                and todoistTask["list"] == "next_actions"
-            ):
-                logger.info(f"Removing task from next actions list.")
-                todoist.delete_task(clickupTask)
-            # Update Clickup task in Todoist
-            elif clickupRequest["event"] in [
+        # Clickup task into todoist next_actions.
+        # Next action status and (high priority, due date < 1 week, no project.)
+        if (
+            clickupRequest["event"] in ["task_updated"]
+            and clickupTask["status"] == "next action"
+            and (
+                clickupTask["priority"] < 3
+                or max_days_diff(clickupTask["due_date"], days=3)
+                or not clickup.is_subtask(clickupTask)
+            )
+        ):
+            outputData["list"] = "next_actions"
+            if not todoistTaskExists:
+                logger.info(f"Adding task to next actions list.")
+                todoistTask = move_task(inputData, outputData, deleteTask=False)
+                clickup.add_todoist_id(clickupTask, todoistTask["todoist_id"])
+            else:
+                logger.info(f"Task is already in next actions list.")
+        elif (
+            clickupRequest["event"] in ["task_updated"]
+            and todoistTaskExists
+            and todoistTask["list"] == "next_actions"
+        ):
+            logger.info(f"Removing task from next actions list.")
+            todoist.delete_task(clickupTask)
+
+        # Update Clickup task in Todoist
+        elif (
+            clickupRequest["event"]
+            in [
                 "task_complete",
                 "task_updated",
                 "task_removed",
-            ]:
-                # Returns False if no task in Todoist
-                todoistTask = modify_task(
-                    inputData, outputData, clickupRequest["event"]
-                )
+            ]
+            and todoistTaskExists
+        ):
+            todoistTask = modify_task(inputData, outputData, clickupRequest["event"])
         return make_response(jsonify({"status": "success"}), 202)
     except Exception as e:
         logging.warning(f"Error in processing clickup webhook: {e}")
